temp commit

+++ examples/housing.mojo
@@ -5,7 +5,6 @@ from basalt import Tensor, TensorShape
 from basalt import Graph, Symbol, OP
 from basalt.utils.datasets import BostonHousing
 from basalt.utils.dataloader import DataLoader
-from basalt.nn.model import Parameters
 
 
 def linear_regression(batch_size: Int, n_inputs: Int, n_outputs: Int) -> Graph:
@@ -20,24 +19,19 @@ def linear_regression(batch_size: Int, n_inputs: Int, n_outputs: Int) -> Graph:
     var loss = nn.MSELoss(g, y_pred, y_true)
     g.loss(loss)
 
-    return g ^
+    return g^
 
 
 def main():
-    # Train Parameters
     comptime batch_size = 32
     comptime num_epochs = 20
     comptime learning_rate = 0.02
 
     comptime graph = linear_regression(batch_size, 13, 1)
 
-    # try: graph.render("operator")
-    # except: print("Could not render graph")
-
     var model = nn.Model[graph]()
     var optim = nn.optim.Adam[graph](model.parameters, lr=learning_rate)
 
-    # Batchwise data loader
     print("Loading data...")
     var train_data: BostonHousing
     try:
@@ -56,10 +50,8 @@ def main():
         var num_batches: Int = 0
         var epoch_loss: Float32 = 0.0
         for batch in training_loader:
-            # Forward pass
             var loss = model.forward(batch.data.copy(), batch.labels.copy()).copy()
 
-            # Backward pass
             optim.zero_grad()
             model.backward()
             optim.step()
@@ -73,14 +65,7 @@ def main():
             "/",
             num_epochs,
             "] \t Avg loss per epoch:",
-            epoch_loss / num_batches,
+            epoch_loss / Float32(num_batches),
         )
 
-    print("Training finished: ", (now() - start) / 1e9, "seconds")
-
-    # print("\n\nInferencing model...\n")
-    # for batch in training_loader:
-    #     var output = model.inference(batch.data)
-
-    #     # Print first (and only output)
-    #     print("Predicted: ", output[0])
+    print("Training finished: ", Float64(now() - start) / 1e9, "seconds")
